requirements.txt
flask==1.1.4
Flask-Admin==1.5.8
flask-restx==0.4.0
Flask-SQLAlchemy==2.5.1
gunicorn==20.1.0
psycopg2-binary==2.8.6
black==21.5b1
flake8===3.9.2
isort==5.8.0
pytest==6.2.4
pytest-cov==2.12.0
pytest-xdist==2.2.1